Separated tests for cpp interface wrappers into separate project
Support for interface wrappers in different environments may lag
behind so it is better that they are in a separate project.
