+++ test/CMakeLists.txt
@@ -5,6 +5,8 @@ add_subdirectory(runpath)
 
 # Collection of available test libraries
 add_subdirectory(testlib)
+add_subdirectory(cpp-utility)
 
 # Collection of available tests
 add_subdirectory(cpp-raw)
+add_subdirectory(cpp-wrapper)
