+++ test/cpp-raw/error_info.cpp
@@ -1,5 +1,6 @@
-#include "os.h"
-#include "catch.hpp"
+#include "../cpp-utility/os.h"
+#include "../cpp-utility/catch.hpp"
+#include "testlib.h"
 
 #ifdef _MSC_VER
