+++ test/cpp-raw/iunknown.cpp
@@ -1,6 +1,6 @@
 
-#include "os.h"
-#include "catch.hpp"
+#include "../cpp-utility/os.h"
+#include "../cpp-utility/catch.hpp"
 
 #include "testlib.h"
