--- test/cpp-raw/msvc/import.h
@@ -1,6 +1,0 @@
-
-#pragma once
-
-#include <windows.h>
-#import "..\testlib\target\debug\test_lib.dll" raw_interfaces_only named_guids
-using namespace TestLib;
