+++ test/cpp-raw/msvc/os.cpp
@@ -1,5 +1,5 @@
 
-#include "../os.h"
+#include "../../cpp-utility/os.h"
 
 #pragma comment(linker, "\"/manifestdependency:name='test_lib' type='win32' version='1.0.0.0'\"" )
