+++ test/cpp-raw/result.cpp
@@ -1,5 +1,5 @@
-#include "os.h"
-#include "catch.hpp"
+#include "../cpp-utility/os.h"
+#include "../cpp-utility/catch.hpp"
 
 #include "testlib.h"
