+++ test/cpp-utility/.gitignore
@@ -1,0 +1,2 @@
+/gcc/test_lib.cpp
+/gcc/test_lib.h
