+++ test/cpp-utility/CMakeLists.txt
@@ -1,4 +1,4 @@
-project (cpp-raw)
+project (cpp-utility)
 
 set (PROJECT_INCLUDE_DIR ${PROJECT_SOURCE_DIR})
 set (PROJECT_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR})
@@ -11,16 +11,18 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 # Define source files.
 set(PROJECT_SRCS
-${PROJECT_SOURCE_DIR}/error_info.cpp
-${PROJECT_SOURCE_DIR}/interface_params.cpp
-${PROJECT_SOURCE_DIR}/iunknown.cpp
-${PROJECT_SOURCE_DIR}/main.cpp
-${PROJECT_SOURCE_DIR}/primitive_tests.cpp
-${PROJECT_SOURCE_DIR}/result.cpp
-${PROJECT_SOURCE_DIR}/return_interfaces.cpp
-${PROJECT_SOURCE_DIR}/stateful.cpp
-${PROJECT_SOURCE_DIR}/strings.cpp
+${PROJECT_SOURCE_DIR}/dummy_interface.cpp
 )
+
+# Specify additional compiler specific helpers.
+set(PROJECT_COMPILER_SPECIFIC_SRC)
+if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU")
+
+set(PROJECT_COMPILER_SPECIFIC_SRC
+    ${PROJECT_SOURCE_DIR}/gcc/os.cpp
+    ${PROJECT_SOURCE_DIR}/gcc/test_lib.cpp
+)
+endif()
 
 # Use the intercom-cli as code generator for test_lib.h/test_lib.cpp
 file(GLOB testlib_sources ${TESTLIB_DIR}/src)
@@ -36,19 +38,7 @@ include_directories("${PROJECT_INCLUDE_DIR}")
 # Ensure "intercom.h" is visible.
 include_directories("${PROJECT_INCLUDE_DIR}/../../intercom-cpp")
 
-add_executable(${PROJECT_NAME} ${PROJECT_SRCS})
-
-# Linking to "runpath" library ensures that RUNPATH is set for ELF executables.
-# All the test libraries built with Rust are copied into this location as well.
-# Forcing the RUNPATH with the help of a dummy library ensures we can
-# dynamically load the libraries with their name alone.
-target_link_libraries(${PROJECT_NAME} runpath)
-
-# Link to static dependencies
-target_link_libraries(${PROJECT_NAME} cpp-utility)
-
-# Add dynamic dependencies
-add_dependencies(${PROJECT_NAME} test_lib)
+add_library(${PROJECT_NAME} STATIC ${PROJECT_SRCS} ${PROJECT_COMPILER_SPECIFIC_SRC})
 
 # Compile specific linking.
 if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU")
