+++ test/cpp-utility/dummy_interface.cpp
