+++ test/cpp-wrapper/CMakeLists.txt
@@ -1,4 +1,4 @@
-project (cpp-raw)
+project (cpp-wrapper)
 
 set (PROJECT_INCLUDE_DIR ${PROJECT_SOURCE_DIR})
 set (PROJECT_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR})
@@ -11,15 +11,9 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 # Define source files.
 set(PROJECT_SRCS
-${PROJECT_SOURCE_DIR}/error_info.cpp
-${PROJECT_SOURCE_DIR}/interface_params.cpp
-${PROJECT_SOURCE_DIR}/iunknown.cpp
 ${PROJECT_SOURCE_DIR}/main.cpp
-${PROJECT_SOURCE_DIR}/primitive_tests.cpp
-${PROJECT_SOURCE_DIR}/result.cpp
-${PROJECT_SOURCE_DIR}/return_interfaces.cpp
-${PROJECT_SOURCE_DIR}/stateful.cpp
-${PROJECT_SOURCE_DIR}/strings.cpp
+${PROJECT_SOURCE_DIR}/exceptions.cpp
+${PROJECT_SOURCE_DIR}/interface_wrappers.cpp
 )
 
 # Use the intercom-cli as code generator for test_lib.h/test_lib.cpp
@@ -36,7 +30,7 @@ include_directories("${PROJECT_INCLUDE_DIR}")
 # Ensure "intercom.h" is visible.
 include_directories("${PROJECT_INCLUDE_DIR}/../../intercom-cpp")
 
-add_executable(${PROJECT_NAME} ${PROJECT_SRCS})
+add_executable(${PROJECT_NAME} ${PROJECT_SRCS} ${PROJECT_COMPILER_SPECIFIC_SRC})
 
 # Linking to "runpath" library ensures that RUNPATH is set for ELF executables.
 # All the test libraries built with Rust are copied into this location as well.
