+++ test/cpp-wrapper/testlib.h
@@ -1,0 +1,6 @@
+
+// Include declarations required to test "testlib".
+
+ #ifdef __GNUC__
+ #include "../cpp-utility/gcc/test_lib.h"
+ #endif
